Changes storage layout
Storage used to be flat, but in anticipation of adding attrachments,
I wanted to keep all data on a thread co-located. So we switched
to a sharded structue based in the id (similar to git, but only one
level).

This commit also adds a path command so you can find the storage for
a trhead given it's id.